+++ Calculadora/src/calculadora/calculadora_teste_geral.java
@@ -8,20 +8,21 @@ public class calculadora_teste_geral {
 		System.out.println("Area do teste de soma");
 		
 		
-		int soma = calc.somar(3,7);
-		System.out.println(soma);
-		
-		
-		 soma = calc.somar(3,0);
-		System.out.println(soma);
-		
-		
-		 soma = calc.somar(0,0);
-		System.out.println(soma);
-		
-		
-		 soma = calc.somar(3,-1);
-		System.out.println(soma);
+		//cenario de teste 1 : Soma de dois valores.
+				int soma = calc.somar(3,7);
+				System.out.println(soma);
+				
+				//cenario de teste 2: Soma de dois valores sendo que um deles zero.
+				 soma = calc.somar(3,0);
+				System.out.println(soma);
+				
+				//cenario de teste 3: Soma de dois valores sendo ambos zero.
+				 soma = calc.somar(0,0);
+				System.out.println(soma);
+				
+				//cenario de teste 4: Soma de dois valores sendo um deles negativo.
+				 soma = calc.somar(3,-1);
+				System.out.println(soma);
 		
 		
 		
@@ -31,18 +32,24 @@ public class calculadora_teste_geral {
 		System.out.println("Area do teste de Subtração");
 		
 		
-		int sub = calc.subtracao(7,3);
-		System.out.println(sub);
-		
-		
-		sub = calc.subtracao(2, 0);
-		System.out.println(sub);
-		
-		sub = calc.subtracao(3, 3);
-		System.out.println(sub);
-		
-		sub = calc.subtracao(5, -5);
-		System.out.println(sub);
+		//cenario de teste 1: Subtração de dois valores.
+				int sub = calc.subtracao(7,3);
+				System.out.println(sub);
+				
+				
+				//cenario de teste 2: Subtração de dois valores sendo um deles igual a zero.
+				sub = calc.subtracao(2, 0);
+				System.out.println(sub);
+				
+				
+				//cenario de teste 3: Subtração de dois numeros iguais.
+				sub = calc.subtracao(3, 3);
+				System.out.println(sub);
+				
+				
+				//cenario de teste 4: Subtração de dois valores sendo um deles negativo
+				sub = calc.subtracao(5, -5);
+				System.out.println(sub);
 		
 		
 		System.out.println("------------------------------------------------------------------");
@@ -50,34 +57,42 @@ public class calculadora_teste_geral {
 		
 		System.out.println("Area do teste de multiplicação");
 		
-		double mult = calc.multiplicacao(8, 7);
-		System.out.println(mult);
-		
-		mult = calc.multiplicacao(9, 0);
-		System.out.println(mult);
-		
-		mult = calc.multiplicacao(0, 0);
-		System.out.println(mult);
-		
-		mult = calc.multiplicacao(8, -7);
-		System.out.println(mult);
+		//cenario de teste 1: Multiplicação de dois valores.
+				int mult = calc.multiplicacao(8, 7);
+				System.out.println(mult);
+				
+				//cenario de teste 2: Multiplicação de dois valores sendo que um deles igual a zero.
+				mult = calc.multiplicacao(9, 0);
+				System.out.println(mult);
+				
+				//cenario de teste 3: Multiplicação de dois numeros sendo os dois iguais a zero.
+				mult = calc.multiplicacao(0, 0);
+				System.out.println(mult);
+				
+				//cenario de teste 4: Multiplicação de dois valores sendo um deles negativo.
+				mult = calc.multiplicacao(8, -7);
+				System.out.println(mult);
 		
 System.out.println("------------------------------------------------------------------");
 		
 		
 		System.out.println("Area do teste de Divisão");
 		
-		double div = calc.divisao(4, 2);
-		System.out.println(div);
-		
-		div = calc.divisao(1, 2);
-		System.out.println(div);
-		
-		div = calc.divisao(2, 2);
-		System.out.println(div);
-		
-		div = calc.divisao(0, 2);
-		System.out.println(div);
+		//cenario de teste 1:Divisão de dois valores.
+				int div = calc.divisao(4, 2);
+				System.out.println(div);
+				
+				//cenario de teste 2: Divisão de dois valores sendo o dividendo menor que o divisor.
+				div = calc.divisao(1, 2);
+				System.out.println(div);
+				
+				//cenario de teste 3: Divisão de dois valroes iguais.
+				div = calc.divisao(2, 2);
+				System.out.println(div);
+				
+				//cenario de teste 4: Divisão de dois valores sendo o dividendo igual a zero.
+				div = calc.divisao(0, 2);
+				System.out.println(div);
 	}
 
 }
